move input and output to new module
